@@ -2,7 +2,8 @@ package cn.ommiao.lockforp.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import cn.ommiao.lockforp.R;
 import cn.ommiao.lockforp.service.LockAccessibilityService;
